+++ changelog.md
@@ -5,6 +5,8 @@
 - Added class `Chats/Channels/Administrator`
 - Added method `assignChannelAdministrator` in `Administrator`
 - Added method `getChannelAdministrator` in `Administrator`
+- Added class `Chats/Channels/Description`
+- Added method `changeChannelDescription` in `Description`
 - **[BREAKING]** Renamed method and class `GraphExecution` to `GraphExecutions`
 - **[BREAKING]** Renamed method and class `Job` to `Jobs`
 - **[BREAKING]** Renamed method and class `Repository` to `Repositories`